删除不必要的文件

+++ tools/videoTools/generator.py
@@ -35,9 +35,7 @@ class ANKVideo():
         self.tagFile="tag.html"
         self.headerSlice="header.mp4"
         self.endingFile="ending.html"
-
         self.innerVideo="talk.mp4"
-
         self.flatReplaceDict=dict()
         self.ffThreadsNum=8
         self.middleTmpFile=dict()
@@ -47,26 +45,33 @@ class ANKVideo():
         self.tagTime=2
 
         self.endingTime=4
+        self.htmlDirName="__ANK_CACHE__"
+        self.preserveTmpFile=False
         pass
     def askForSetting(self):
         inString=""
         inString=input("input payload video file name  (default talk.mp4)>")
         self.innerVideo=self.innerVideo if len(inString)==0 else inString
+        # self.innerVideo=os.path.join(self.htmlDirName,self.innerVideo)
 
         inString=input("input cover page (default cover.html)>")
         self.coverFile=self.coverFile if len(inString)==0 else inString
+        # self.coverFile=os.path.join(self.htmlDirName,self.coverFile)
+
         inString=input("input profile page (default profile.html)>")
-        self.profileFile=self.profileFile if len(inString)==0 else inString  
+        self.profileFile=self.profileFile if len(inString)==0 else inString
+        # self.profileFile=os.path.join(self.htmlDirName,self.profileFile)
 
         inString=input("input ending page (default ending.html)>")
         self.endingFile=self.endingFile if len(inString)==0 else inString
+        # self.endingFile=os.path.join(self.htmlDirName,self.endingFile)
 
         inString=input("input tag page (default tag.html)>")
-        self.tagFile=self.tagFile if len(inString)==0 else inString 
+        self.tagFile=self.tagFile if len(inString)==0 else inString
+        # self.tagFile=os.path.join(self.htmlDirName,self.tagFile)
+
         inString=input("input header video name (default header.mp4)>")
         self.headerSlice=self.headerSlice if len(inString)==0 else inString
-
-
 
         inString=input("input output video name (default output.mp4)>")
         self.outputFileName=self.outputFileName if len(inString)==0 else inString
@@ -84,6 +89,8 @@ class ANKVideo():
         inString=input("input ending time (default 4)>")
         self.endingTime=self.endingTime if len(inString)==0 else inString
 
+        inString=input("preserve tmp file ? input any char means yes(default No)>")
+        self.preserveTmpFile=self.preserveTmpFile if len(inString)==0 else True
     def _readFileAsDict(self,fileName):
         d=None
         with open(fileName, encoding='utf-8') as f:
@@ -91,8 +98,12 @@ class ANKVideo():
             d = json.loads(lines)
             f.close()
         return d
-    def _readFileAsString(self,fileName):
+    def _readFileAsString(self,fileName,preDir=""):
         lines=None
+        if preDir=="":
+            pass
+        else:
+            fileName=os.path.join(preDir,fileName)
         with open(fileName, encoding='utf-8') as f:
             lines = f.read()
             f.close()
@@ -100,20 +111,17 @@ class ANKVideo():
 
     def build(self):
         self.conf=self._readFileAsDict(self.confFileName)
-        self.flatReplaceDict=flatten(self.conf)
         #handle cover
         coverOutHtml="tmp."+self.coverFile
-        coverInHtmlString=self._readFileAsString(self.coverFile)
+        coverInHtmlString=self._readFileAsString(self.coverFile,self.htmlDirName)
         
-        # for k,v in self.flatReplaceDict.items():
-        #     coverInHtmlString=coverInHtmlString.replace(k,str(v))
         coverInHtmlString=jinja2.Template(coverInHtmlString).render(self.conf)
         writeStringToFile(coverInHtmlString,coverOutHtml)
         self.middleTmpFile["TMP.COVER"]=coverOutHtml
         self.middleTmpFile["TMP.COVER.PNG"]=coverOutHtml+".png"
 
         profileOutHtml="tmp."+self.profileFile
-        profileInHtmlString=self._readFileAsString(self.profileFile)
+        profileInHtmlString=self._readFileAsString(self.profileFile,self.htmlDirName)
         profileInHtmlString=jinja2.Template(profileInHtmlString).render(self.conf)
         writeStringToFile(profileInHtmlString,profileOutHtml)
 
@@ -122,7 +130,7 @@ class ANKVideo():
         self.middleTmpFile["TMP.PROFILE.PNG"]=profileOutHtml+".png"
 
         tagOutHtml="tmp."+self.tagFile
-        tagInHtmlString=self._readFileAsString(self.tagFile)
+        tagInHtmlString=self._readFileAsString(self.tagFile,self.htmlDirName)
         tagInHtmlString=jinja2.Template(tagInHtmlString).render(self.conf)
         writeStringToFile(tagInHtmlString,tagOutHtml)
         self.middleTmpFile["TMP.TAG"]=tagOutHtml
@@ -130,7 +138,7 @@ class ANKVideo():
 
 
         endingOutHtml="tmp."+self.endingFile
-        endingInHtmlString=self._readFileAsString(self.endingFile)
+        endingInHtmlString=self._readFileAsString(self.endingFile,self.htmlDirName)
         endingInHtmlString=jinja2.Template(endingInHtmlString).render(self.conf)
         writeStringToFile(endingInHtmlString,endingOutHtml)
         self.middleTmpFile["TMP.END"]=endingOutHtml
@@ -153,59 +161,6 @@ class ANKVideo():
         pTag.wait()
         endTag.wait()
 
-        # self.middleTmpFile["TMP.COVER.PNG.MP4"]=self.middleTmpFile["TMP.COVER.PNG"]+".mp4"
-        # self.middleTmpFile["TMP.PROFILE.PNG.MP4"]=self.middleTmpFile["TMP.PROFILE.PNG"]+".mp4"
-        # self.middleTmpFile["TMP.TAG.PNG.MP4"]=self.middleTmpFile["TMP.TAG.PNG"]+".mp4"
-        # self.middleTmpFile["TMP.END.PNG.MP4"]=self.middleTmpFile["TMP.END.PNG"]+".mp4"
-
-        # pngToMp4CommandShell="ffmpeg -r 25 -loop 1 -i "+ self.middleTmpFile["TMP.COVER.PNG"]+ " -pix_fmt yuv420p -vcodec libx264 -b:v 600k -r:v 25 -preset medium -crf 25 -s 1920x1080 -vframes 250  -r 25 -t "+str(self.coverTime)+"  -threads 8 "+self.middleTmpFile["TMP.COVER.PNG.MP4"]
-
-        # cvtCoverP=subprocess.Popen(pngToMp4CommandShell,shell=True)
-
-
-        # pngToMp4CommandShell="ffmpeg -r 25 -loop 1 -i "+ self.middleTmpFile["TMP.PROFILE.PNG"]+ " -pix_fmt yuv420p -vcodec libx264 -b:v 600k -r:v 25 -preset medium -crf 25 -s 1920x1080 -vframes 250 -r 25 -t "+str(self.profileTime)+"  -threads 8 "+self.middleTmpFile["TMP.PROFILE.PNG.MP4"]
-
-        # cvtProfileP=subprocess.Popen(pngToMp4CommandShell,shell=True)
-
-
-        # pngToMp4CommandShell="ffmpeg -r 25 -loop 1 -i "+ self.middleTmpFile["TMP.TAG.PNG"]+ " -pix_fmt yuv420p -vcodec libx264 -b:v 600k -r:v 25 -preset medium -crf 25 -s 1920x1080 -vframes 250 -r 25 -t "+str(self.tagTime)+"  -threads 8 "+self.middleTmpFile["TMP.TAG.PNG.MP4"]
-
-        # cvtTagP=subprocess.Popen(pngToMp4CommandShell,shell=True)
-
-
-        # pngToMp4CommandShell="ffmpeg -r 25 -loop 1 -i "+ self.middleTmpFile["TMP.END.PNG"]+ " -pix_fmt yuv420p -vcodec libx264 -b:v 600k -r:v 25 -preset medium -crf 25 -s 1920x1080 -vframes 250 -r 25 -t "+str(self.endingTime)+"  -threads 8 "+self.middleTmpFile["TMP.END.PNG.MP4"]
-
-        # cvtEnd=subprocess.Popen(pngToMp4CommandShell,shell=True)
-
-        # cvtCoverP.wait()
-        # cvtProfileP.wait()
-        # cvtTagP.wait()
-        # cvtEnd.wait()
-
-        # files=[]
-
-        # files.append("file '"+self.middleTmpFile["TMP.COVER.PNG.MP4"]+"'")
-        # files.append("file '"+self.middleTmpFile["TMP.PROFILE.PNG.MP4"]+"'")
-        # files.append("file '"+self.middleTmpFile["TMP.TAG.PNG.MP4"]+"'")
-        # if len(self.innerVideo)==0:
-        #     pass
-        # else:
-        #     files.append("file '"+self.innerVideo+"'")
-
-        # files.append("file '"+self.middleTmpFile["TMP.END.PNG.MP4"]+"'")
-
-
-        # finalString=str.join("\n",files)
-        
-        # writeStringToFile(finalString,"tmp.list.txt")
-
-
-        # joinVideoCommandShell="ffmpeg -f concat -i tmp.list.txt -c copy -acodec copy  "+self.outputFileName
-
-        # joinVideoP=subprocess.Popen(joinVideoCommandShell,shell=True)
-
-
-
         joinVideoCommandShell='ffmpeg \
 -loop 1 -probesize 10MB -framerate 24 -t {} -i {} \
 -loop 1 -probesize 10MB -framerate 24 -t {} -i {} \
@@ -216,19 +171,19 @@ class ANKVideo():
 -filter_complex " [0:v][5:a][1:v][5:a][2:v][5:a][3:v][3:a][4:v][5:a] concat=n=5:v=1:a=1 [v][a]" \
 -map "[v]" -map "[a]" {}'.format(self.coverTime,self.middleTmpFile["TMP.COVER.PNG"],self.profileTime,self.middleTmpFile["TMP.PROFILE.PNG"],self.tagTime,self.middleTmpFile["TMP.TAG.PNG"],self.innerVideo,self.endingTime,self.middleTmpFile["TMP.END.PNG"],self.outputFileName)
 # "+self.outputFileName
-
+        print(joinVideoCommandShell)
         joinVideoP=subprocess.Popen(joinVideoCommandShell,shell=True)
-
-
-
         joinVideoP.wait()
 
-        try:
-            for k,v in self.middleTmpFile.items():
-                os.remove(v)
-                pass
-        except Exception:
+        if self.preserveTmpFile:
             pass
+        else:
+            try:
+                for k,v in self.middleTmpFile.items():
+                    os.remove(v)
+                    pass
+            except Exception:
+                pass
         # print(str(self.middleTmpFile))
